Added to House
Added the navigation to get into the House, Backyard, and back into the DungeonRoom. Found a bug where the player has no choice but to exit dungeon room if the door is unlocked and they are standing in front of it. Ignores all other options. Look into fixing this, and having slightly different options to match the re-entry into dungeonRoom. Make sure future rooms have more concise and easier to follow logic.